CQ-4279630 - [FluidXP] Add Sites FluidXP sanity tests in ESP
Code cleanup
CQ-4279630 - [FluidXP] Add Sites FluidXP sanity tests in ESP

Create one class to contain the basic CRD tests (performance
improvement)
